layout

+++ database/migrations/2024_05_06_024148_create_working_times_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->unsignedBigInteger('employee_id');
             $table->dateTime('start_time');
             $table->dateTime('end_time');
-            $table->integer('total') ;
+            $table->float('total') ;
             $table->timestamps();
 
             $table->foreign('employee_id')->references('id')->on('users')->onDelete('cascade');
